fix to opening files with docker-tramp (#40)
* fix to opening files with docker-tramp

Use a consistent docker container name between running the container
and the uri->path mapping.

Capture the suffix at the time of registering the client. The same
suffix that the container is started with is used when mapping file
uri to paths.

* verify suffix type is a number before it is incremented

Do not increment `lsp-docker-container-name-suffix` if it has been
customized to not contain a number.